+++ .gitmodules
@@ -2,7 +2,7 @@
 	path = src/core/libdeno/build
 	url = https://github.com/denoland/chromium_build.git
 [submodule "deno_std"]
-	path = src/js/deps/https/deno.land/std
+	path = src/std
 	url = https://github.com/denoland/deno_std.git
 [submodule "deno_third_party"]
 	path = third_party
