+++ tools/benchmark.py
@@ -217,8 +217,8 @@ def run_http(build_dir, new_data):
 
 def bundle_benchmark(deno_exe):
     bundles = {
-        "file_server": "./src/js/deps/https/deno.land/std/http/file_server.ts",
-        "gist": "./src/js/deps/https/deno.land/std/examples/gist.ts",
+        "file_server": "./src/std/http/file_server.ts",
+        "gist": "./src/std/examples/gist.ts",
     }
 
     sizes = {}
