+++ tools/deno_http_proxy.ts
@@ -1,8 +1,5 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import {
-  serve,
-  ServerRequest
-} from "../src/js/deps/https/deno.land/std/http/server.ts";
+import { serve, ServerRequest } from "../src/std/http/server.ts";
 
 const addr = Deno.args[1] || "127.0.0.1:4500";
 const originAddr = Deno.args[2] || "127.0.0.1:4501";
